Fixed example Python code
On the Get Started page, the example Python code was slightly wrong. There should be no `/api/v1/` in the URL passed to `BigchainDB()`. This PR fixes that error. I made a similar fix on the IPDB Developer Portal home page a while back. I just noticed the same error was present on the Get Started page.