Update getstarted.yml: remove transaction-creation, call-to-action to IPDB
The first part on transaction-creation ("Send your first transaction" etc) is broken because BigchainDB is no longer hosting an instance of BigchainDB software. So, remove it.

There are two network options listed. Before, one was using the BigchainDB-hosted network, but that's not running. So change the text to point to the IPDB resources page, which includes pointers to IPDB networks. Also not that this IPDB page has its own "Send your first transaction" information:)